@@ -22,11 +22,11 @@ class $FloorAppDatabase {
 class _$AppDatabaseBuilder {
   _$AppDatabaseBuilder(this.name);
 
-  final String name;
+  final String? name;
 
   final List<Migration> _migrations = [];
 
-  Callback _callback;
+  Callback? _callback;
 
   /// Adds migrations to the builder.
   _$AppDatabaseBuilder addMigrations(List<Migration> migrations) {
@@ -43,7 +43,7 @@ class _$AppDatabaseBuilder {
   /// Creates the database and initializes it.
   Future<AppDatabase> build() async {
     final path = name != null
-        ? await sqfliteDatabaseFactory.getDatabasePath(name)
+        ? await sqfliteDatabaseFactory.getDatabasePath(name!)
         : ':memory:';
     final database = _$AppDatabase();
     database.database = await database.open(
@@ -56,32 +56,33 @@ class _$AppDatabaseBuilder {
 }
 
 class _$AppDatabase extends AppDatabase {
-  _$AppDatabase([StreamController<String> listener]) {
+  _$AppDatabase([StreamController<String>? listener]) {
     changeListener = listener ?? StreamController<String>.broadcast();
   }
 
-  UserDao _userDaoInstance;
+  UserDao? _userDaoInstance;
 
-  AccountDao _accountDaoInstance;
+  AccountDao? _accountDaoInstance;
 
-  CurrencyDao _currencyDaoInstance;
+  CurrencyDao? _currencyDaoInstance;
 
-  TransactionDao _transactionDaoInstance;
+  TransactionDao? _transactionDaoInstance;
 
-  NetworkDao _networkDaoInstance;
+  NetworkDao? _networkDaoInstance;
 
-  AccountCurrencyDao _accountCurrencyDaoInstance;
+  AccountCurrencyDao? _accountCurrencyDaoInstance;
 
-  UtxoDao _utxoDaoInstance;
+  UtxoDao? _utxoDaoInstance;
 
-  ExchangeRateDao _exchangeRateDaoInstance;
+  ExchangeRateDao? _exchangeRateDaoInstance;
 
   Future<sqflite.Database> open(String path, List<Migration> migrations,
-      [Callback callback]) async {
+      [Callback? callback]) async {
     final databaseOptions = sqflite.OpenDatabaseOptions(
       version: 1,
       onConfigure: (database) async {
         await database.execute('PRAGMA foreign_keys = ON');
+        await callback?.onConfigure?.call(database);
       },
       onOpen: (database) async {
         await callback?.onOpen?.call(database);
@@ -94,21 +95,21 @@ class _$AppDatabase extends AppDatabase {
       },
       onCreate: (database, version) async {
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `User` (`user_id` TEXT, `keystore` TEXT, `third_party_id` TEXT, `install_id` TEXT, `timestamp` INTEGER, `backup_status` INTEGER NOT NULL, PRIMARY KEY (`user_id`))');
+            'CREATE TABLE IF NOT EXISTS `User` (`user_id` TEXT NOT NULL, `keystore` TEXT NOT NULL, `third_party_id` TEXT NOT NULL, `install_id` TEXT NOT NULL, `timestamp` INTEGER NOT NULL, `last_sync_time:` INTEGER NOT NULL, PRIMARY KEY (`user_id`))');
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `Account` (`account_id` TEXT, `user_id` TEXT, `network_id` TEXT, `account_index` INTEGER, FOREIGN KEY (`user_id`) REFERENCES `User` (`user_id`) ON UPDATE NO ACTION ON DELETE CASCADE, PRIMARY KEY (`account_id`))');
+            'CREATE TABLE IF NOT EXISTS `Account` (`account_id` TEXT NOT NULL, `user_id` TEXT NOT NULL, `network_id` TEXT NOT NULL, `account_index` INTEGER NOT NULL, FOREIGN KEY (`user_id`) REFERENCES `User` (`user_id`) ON UPDATE NO ACTION ON DELETE CASCADE, PRIMARY KEY (`account_id`))');
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `Currency` (`currency_id` TEXT, `name` TEXT, `description` TEXT, `symbol` TEXT, `decimals` INTEGER, `address` TEXT, `type` TEXT, `total_supply` TEXT, `contract` TEXT, `image` TEXT, PRIMARY KEY (`currency_id`))');
+            'CREATE TABLE IF NOT EXISTS `Currency` (`currency_id` TEXT NOT NULL, `name` TEXT NOT NULL, `description` TEXT, `symbol` TEXT NOT NULL, `decimals` INTEGER NOT NULL, `address` TEXT NOT NULL, `type` TEXT NOT NULL, `total_supply` TEXT, `contract` TEXT, `image` TEXT NOT NULL, PRIMARY KEY (`currency_id`))');
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `_Transaction` (`transaction_id` TEXT, `accountcurrency_id` TEXT, `tx_id` TEXT, `source_address` TEXT, `destinction_address` TEXT, `timestamp` INTEGER, `confirmation` INTEGER, `gas_price` TEXT, `gas_used` INTEGER, `fee` TEXT NOT NULL, `note` TEXT, `status` TEXT, `direction` TEXT, `amount` TEXT, PRIMARY KEY (`transaction_id`))');
+            'CREATE TABLE IF NOT EXISTS `_Transaction` (`transaction_id` TEXT NOT NULL, `accountcurrency_id` TEXT NOT NULL, `tx_id` TEXT NOT NULL, `source_address` TEXT NOT NULL, `destinction_address` TEXT NOT NULL, `timestamp` INTEGER NOT NULL, `confirmation` INTEGER NOT NULL, `gas_price` TEXT, `gas_used` INTEGER, `fee` TEXT NOT NULL, `note` TEXT, `status` TEXT NOT NULL, `direction` TEXT NOT NULL, `amount` TEXT NOT NULL, PRIMARY KEY (`transaction_id`))');
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `Network` (`network_id` TEXT, `network` TEXT NOT NULL, `coin_type` INTEGER, `publish` INTEGER, `chain_id` INTEGER, PRIMARY KEY (`network_id`))');
+            'CREATE TABLE IF NOT EXISTS `Network` (`network_id` TEXT NOT NULL, `network` TEXT NOT NULL, `coin_type` INTEGER NOT NULL, `publish` INTEGER NOT NULL, `chain_id` INTEGER NOT NULL, PRIMARY KEY (`network_id`))');
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `AccountCurrency` (`accountcurrency_id` TEXT NOT NULL, `account_id` TEXT, `currency_id` TEXT, `balance` TEXT, `number_of_used_external_key` INTEGER, `number_of_used_internal_key` INTEGER, `last_sync_time` INTEGER, FOREIGN KEY (`account_id`) REFERENCES `Account` (`account_id`) ON UPDATE NO ACTION ON DELETE CASCADE, FOREIGN KEY (`currency_id`) REFERENCES `Currency` (`currency_id`) ON UPDATE NO ACTION ON DELETE CASCADE, PRIMARY KEY (`accountcurrency_id`))');
+            'CREATE TABLE IF NOT EXISTS `AccountCurrency` (`accountcurrency_id` TEXT NOT NULL, `account_id` TEXT NOT NULL, `currency_id` TEXT NOT NULL, `balance` TEXT NOT NULL, `number_of_used_external_key` INTEGER NOT NULL, `number_of_used_internal_key` INTEGER NOT NULL, `last_sync_time` INTEGER NOT NULL, FOREIGN KEY (`account_id`) REFERENCES `Account` (`account_id`) ON UPDATE NO ACTION ON DELETE CASCADE, FOREIGN KEY (`currency_id`) REFERENCES `Currency` (`currency_id`) ON UPDATE NO ACTION ON DELETE CASCADE, PRIMARY KEY (`accountcurrency_id`))');
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `Utxo` (`utxo_id` TEXT NOT NULL, `accountcurrency_id` TEXT, `tx_id` TEXT, `vout` INTEGER, `type` TEXT, `amount` TEXT, `chain_index` INTEGER, `key_index` INTEGER, `script` TEXT, `timestamp` INTEGER, `locked` INTEGER, `sequence` INTEGER, `address` TEXT, FOREIGN KEY (`accountcurrency_id`) REFERENCES `AccountCurrency` (`accountcurrency_id`) ON UPDATE NO ACTION ON DELETE CASCADE, PRIMARY KEY (`utxo_id`))');
+            'CREATE TABLE IF NOT EXISTS `Utxo` (`utxo_id` TEXT NOT NULL, `accountcurrency_id` TEXT NOT NULL, `tx_id` TEXT NOT NULL, `vout` INTEGER NOT NULL, `type` TEXT NOT NULL, `amount` TEXT NOT NULL, `chain_index` INTEGER NOT NULL, `key_index` INTEGER NOT NULL, `script` TEXT NOT NULL, `timestamp` INTEGER NOT NULL, `locked` INTEGER NOT NULL, `sequence` INTEGER NOT NULL, `address` TEXT NOT NULL, FOREIGN KEY (`accountcurrency_id`) REFERENCES `AccountCurrency` (`accountcurrency_id`) ON UPDATE NO ACTION ON DELETE CASCADE, PRIMARY KEY (`utxo_id`))');
         await database.execute(
-            'CREATE TABLE IF NOT EXISTS `ExchangeRate` (`exchange_rate_id` TEXT, `name` TEXT, `rate` TEXT, `lastSyncTime` INTEGER, `type` TEXT, PRIMARY KEY (`exchange_rate_id`))');
+            'CREATE TABLE IF NOT EXISTS `ExchangeRate` (`exchange_rate_id` TEXT NOT NULL, `name` TEXT NOT NULL, `rate` TEXT NOT NULL, `lastSyncTime` INTEGER NOT NULL, `type` TEXT NOT NULL, PRIMARY KEY (`exchange_rate_id`))');
 
         await database.execute(
             '''CREATE VIEW IF NOT EXISTS `JoinCurrency` AS SELECT * FROM AccountCurrency INNER JOIN Currency ON AccountCurrency.currency_id = Currency.currency_id INNER JOIN Account ON AccountCurrency.account_id = Account.account_id INNER JOIN Network ON Account.network_id = Network.network_id''');
@@ -173,37 +174,37 @@ class _$UserDao extends UserDao {
         _userEntityInsertionAdapter = InsertionAdapter(
             database,
             'User',
-            (UserEntity item) => <String, dynamic>{
+            (UserEntity item) => <String, Object?>{
                   'user_id': item.userId,
                   'keystore': item.keystore,
                   'third_party_id': item.thirdPartyId,
                   'install_id': item.installId,
                   'timestamp': item.timestamp,
-                  'backup_status': item.backupStatus ? 1 : 0
+                  'last_sync_time:': item.lastSyncTime
                 }),
         _userEntityUpdateAdapter = UpdateAdapter(
             database,
             'User',
             ['user_id'],
-            (UserEntity item) => <String, dynamic>{
+            (UserEntity item) => <String, Object?>{
                   'user_id': item.userId,
                   'keystore': item.keystore,
                   'third_party_id': item.thirdPartyId,
                   'install_id': item.installId,
                   'timestamp': item.timestamp,
-                  'backup_status': item.backupStatus ? 1 : 0
+                  'last_sync_time:': item.lastSyncTime
                 }),
         _userEntityDeletionAdapter = DeletionAdapter(
             database,
             'User',
             ['user_id'],
-            (UserEntity item) => <String, dynamic>{
+            (UserEntity item) => <String, Object?>{
                   'user_id': item.userId,
                   'keystore': item.keystore,
                   'third_party_id': item.thirdPartyId,
                   'install_id': item.installId,
                   'timestamp': item.timestamp,
-                  'backup_status': item.backupStatus ? 1 : 0
+                  'last_sync_time:': item.lastSyncTime
                 });
 
   final sqflite.DatabaseExecutor database;
@@ -219,15 +220,15 @@ class _$UserDao extends UserDao {
   final DeletionAdapter<UserEntity> _userEntityDeletionAdapter;
 
   @override
-  Future<UserEntity> findUser() async {
+  Future<UserEntity?> findUser() async {
     return _queryAdapter.query('SELECT * FROM User limit 1',
-        mapper: (Map<String, dynamic> row) => UserEntity(
+        mapper: (Map<String, Object?> row) => UserEntity(
             row['user_id'] as String,
             row['keystore'] as String,
             row['third_party_id'] as String,
             row['install_id'] as String,
             row['timestamp'] as int,
-            (row['backup_status'] as int) != 0));
+            row['last_sync_time:'] as int));
   }
 
   @override
@@ -252,7 +253,7 @@ class _$AccountDao extends AccountDao {
         _accountEntityInsertionAdapter = InsertionAdapter(
             database,
             'Account',
-            (AccountEntity item) => <String, dynamic>{
+            (AccountEntity item) => <String, Object?>{
                   'account_id': item.accountId,
                   'user_id': item.userId,
                   'network_id': item.networkId,
@@ -268,9 +269,9 @@ class _$AccountDao extends AccountDao {
   final InsertionAdapter<AccountEntity> _accountEntityInsertionAdapter;
 
   @override
-  Future<List<AccountEntity>> findAllAccounts() async {
+  Future<List<AccountEntity>?> findAllAccounts() async {
     return _queryAdapter.queryList('SELECT * FROM Account',
-        mapper: (Map<String, dynamic> row) => AccountEntity(
+        mapper: (Map<String, Object?> row) => AccountEntity(
             accountId: row['account_id'] as String,
             userId: row['user_id'] as String,
             networkId: row['network_id'] as String,
@@ -278,15 +279,15 @@ class _$AccountDao extends AccountDao {
   }
 
   @override
-  Future<AccountEntity> findAccount(String id) async {
+  Future<AccountEntity?> findAccount(String id) async {
     return _queryAdapter.query(
-        'SELECT * FROM Account WHERE account_id = ? LIMIT 1',
-        arguments: <dynamic>[id],
-        mapper: (Map<String, dynamic> row) => AccountEntity(
+        'SELECT * FROM Account WHERE account_id = ?1 LIMIT 1',
+        mapper: (Map<String, Object?> row) => AccountEntity(
             accountId: row['account_id'] as String,
             userId: row['user_id'] as String,
             networkId: row['network_id'] as String,
-            accountIndex: row['account_index'] as int));
+            accountIndex: row['account_index'] as int),
+        arguments: [id]);
   }
 
   @override
@@ -308,7 +309,7 @@ class _$CurrencyDao extends CurrencyDao {
         _currencyEntityInsertionAdapter = InsertionAdapter(
             database,
             'Currency',
-            (CurrencyEntity item) => <String, dynamic>{
+            (CurrencyEntity item) => <String, Object?>{
                   'currency_id': item.currencyId,
                   'name': item.name,
                   'description': item.description,
@@ -330,31 +331,31 @@ class _$CurrencyDao extends CurrencyDao {
   final InsertionAdapter<CurrencyEntity> _currencyEntityInsertionAdapter;
 
   @override
-  Future<List<CurrencyEntity>> findAllCurrencies() async {
+  Future<List<CurrencyEntity>?> findAllCurrencies() async {
     return _queryAdapter.queryList('SELECT * FROM Currency',
-        mapper: (Map<String, dynamic> row) => CurrencyEntity(
+        mapper: (Map<String, Object?> row) => CurrencyEntity(
             currencyId: row['currency_id'] as String,
             name: row['name'] as String,
             symbol: row['symbol'] as String,
-            description: row['description'] as String,
+            description: row['description'] as String?,
             address: row['address'] as String,
-            contract: row['contract'] as String,
+            contract: row['contract'] as String?,
             decimals: row['decimals'] as int,
-            totalSupply: row['total_supply'] as String,
+            totalSupply: row['total_supply'] as String?,
             type: row['type'] as String,
             image: row['image'] as String));
   }
 
   @override
-  Future<List<CurrencyWithAccountId>> findAllCurrenciesByAccountId(
+  Future<List<CurrencyWithAccountId>?> findAllCurrenciesByAccountId(
       String id) async {
     return _queryAdapter.queryList(
-        'SELECT * FROM CurrencyWithAccountId where account_id = ?',
-        arguments: <dynamic>[id],
-        mapper: (Map<String, dynamic> row) => CurrencyWithAccountId(
+        'SELECT * FROM CurrencyWithAccountId where account_id = ?1',
+        mapper: (Map<String, Object?> row) => CurrencyWithAccountId(
             currencyId: row['currency_id'] as String,
             accountId: row['account_id'] as String,
-            symbol: row['symbol'] as String));
+            symbol: row['symbol'] as String),
+        arguments: [id]);
   }
 
   @override
@@ -376,7 +377,7 @@ class _$TransactionDao extends TransactionDao {
         _transactionEntityInsertionAdapter = InsertionAdapter(
             database,
             '_Transaction',
-            (TransactionEntity item) => <String, dynamic>{
+            (TransactionEntity item) => <String, Object?>{
                   'transaction_id': item.transactionId,
                   'accountcurrency_id': item.accountcurrencyId,
                   'tx_id': item.txId,
@@ -396,7 +397,7 @@ class _$TransactionDao extends TransactionDao {
             database,
             '_Transaction',
             ['transaction_id'],
-            (TransactionEntity item) => <String, dynamic>{
+            (TransactionEntity item) => <String, Object?>{
                   'transaction_id': item.transactionId,
                   'accountcurrency_id': item.accountcurrencyId,
                   'tx_id': item.txId,
@@ -416,7 +417,7 @@ class _$TransactionDao extends TransactionDao {
             database,
             '_Transaction',
             ['transaction_id'],
-            (TransactionEntity item) => <String, dynamic>{
+            (TransactionEntity item) => <String, Object?>{
                   'transaction_id': item.transactionId,
                   'accountcurrency_id': item.accountcurrencyId,
                   'tx_id': item.txId,
@@ -446,18 +447,18 @@ class _$TransactionDao extends TransactionDao {
   final DeletionAdapter<TransactionEntity> _transactionEntityDeletionAdapter;
 
   @override
-  Future<List<TransactionEntity>> findAllTransactions() async {
+  Future<List<TransactionEntity>?> findAllTransactions() async {
     return _queryAdapter.queryList('SELECT * FROM _Transaction',
-        mapper: (Map<String, dynamic> row) => TransactionEntity(
+        mapper: (Map<String, Object?> row) => TransactionEntity(
             transactionId: row['transaction_id'] as String,
             accountcurrencyId: row['accountcurrency_id'] as String,
             txId: row['tx_id'] as String,
             confirmation: row['confirmation'] as int,
             sourceAddress: row['source_address'] as String,
             destinctionAddress: row['destinction_address'] as String,
-            gasPrice: row['gas_price'] as String,
-            gasUsed: row['gas_used'] as int,
-            note: row['note'] as String,
+            gasPrice: row['gas_price'] as String?,
+            gasUsed: row['gas_used'] as int?,
+            note: row['note'] as String?,
             fee: row['fee'] as String,
             status: row['status'] as String,
             timestamp: row['timestamp'] as int,
@@ -466,47 +467,47 @@ class _$TransactionDao extends TransactionDao {
   }
 
   @override
-  Future<List<TransactionEntity>> findAllTransactionsById(String id) async {
+  Future<List<TransactionEntity>?> findAllTransactionsById(String id) async {
     return _queryAdapter.queryList(
-        'SELECT * FROM _Transaction WHERE _Transaction.accountcurrency_id = ?',
-        arguments: <dynamic>[id],
-        mapper: (Map<String, dynamic> row) => TransactionEntity(
+        'SELECT * FROM _Transaction WHERE _Transaction.accountcurrency_id = ?1',
+        mapper: (Map<String, Object?> row) => TransactionEntity(
             transactionId: row['transaction_id'] as String,
             accountcurrencyId: row['accountcurrency_id'] as String,
             txId: row['tx_id'] as String,
             confirmation: row['confirmation'] as int,
             sourceAddress: row['source_address'] as String,
             destinctionAddress: row['destinction_address'] as String,
-            gasPrice: row['gas_price'] as String,
-            gasUsed: row['gas_used'] as int,
-            note: row['note'] as String,
+            gasPrice: row['gas_price'] as String?,
+            gasUsed: row['gas_used'] as int?,
+            note: row['note'] as String?,
             fee: row['fee'] as String,
             status: row['status'] as String,
             timestamp: row['timestamp'] as int,
             direction: row['direction'] as String,
-            amount: row['amount'] as String));
+            amount: row['amount'] as String),
+        arguments: [id]);
   }
 
   @override
-  Future<TransactionEntity> findTransactionsByTxId(String id) async {
+  Future<TransactionEntity?> findTransactionsByTxId(String id) async {
     return _queryAdapter.query(
-        'SELECT * FROM _Transaction WHERE _Transaction.tx_id = ? limit 1',
-        arguments: <dynamic>[id],
-        mapper: (Map<String, dynamic> row) => TransactionEntity(
+        'SELECT * FROM _Transaction WHERE _Transaction.tx_id = ?1 limit 1',
+        mapper: (Map<String, Object?> row) => TransactionEntity(
             transactionId: row['transaction_id'] as String,
             accountcurrencyId: row['accountcurrency_id'] as String,
             txId: row['tx_id'] as String,
             confirmation: row['confirmation'] as int,
             sourceAddress: row['source_address'] as String,
             destinctionAddress: row['destinction_address'] as String,
-            gasPrice: row['gas_price'] as String,
-            gasUsed: row['gas_used'] as int,
-            note: row['note'] as String,
+            gasPrice: row['gas_price'] as String?,
+            gasUsed: row['gas_used'] as int?,
+            note: row['note'] as String?,
             fee: row['fee'] as String,
             status: row['status'] as String,
             timestamp: row['timestamp'] as int,
             direction: row['direction'] as String,
-            amount: row['amount'] as String));
+            amount: row['amount'] as String),
+        arguments: [id]);
   }
 
   @override
@@ -538,12 +539,11 @@ class _$NetworkDao extends NetworkDao {
         _networkEntityInsertionAdapter = InsertionAdapter(
             database,
             'Network',
-            (NetworkEntity item) => <String, dynamic>{
+            (NetworkEntity item) => <String, Object?>{
                   'network_id': item.networkId,
                   'network': item.network,
                   'coin_type': item.coinType,
-                  'publish':
-                      item.publish == null ? null : (item.publish ? 1 : 0),
+                  'publish': item.publish ? 1 : 0,
                   'chain_id': item.chainId
                 });
 
@@ -556,14 +556,13 @@ class _$NetworkDao extends NetworkDao {
   final InsertionAdapter<NetworkEntity> _networkEntityInsertionAdapter;
 
   @override
-  Future<List<NetworkEntity>> findAllNetworks() async {
+  Future<List<NetworkEntity>?> findAllNetworks() async {
     return _queryAdapter.queryList('SELECT * FROM Network',
-        mapper: (Map<String, dynamic> row) => NetworkEntity(
+        mapper: (Map<String, Object?> row) => NetworkEntity(
             networkId: row['network_id'] as String,
             network: row['network'] as String,
             coinType: row['coin_type'] as int,
-            publish:
-                row['publish'] == null ? null : (row['publish'] as int) != 0,
+            publish: (row['publish'] as int) != 0,
             chainId: row['chain_id'] as int));
   }
 
@@ -580,7 +579,7 @@ class _$AccountCurrencyDao extends AccountCurrencyDao {
         _accountCurrencyEntityInsertionAdapter = InsertionAdapter(
             database,
             'AccountCurrency',
-            (AccountCurrencyEntity item) => <String, dynamic>{
+            (AccountCurrencyEntity item) => <String, Object?>{
                   'accountcurrency_id': item.accountcurrencyId,
                   'account_id': item.accountId,
                   'currency_id': item.currencyId,
@@ -600,9 +599,9 @@ class _$AccountCurrencyDao extends AccountCurrencyDao {
       _accountCurrencyEntityInsertionAdapter;
 
   @override
-  Future<List<AccountCurrencyEntity>> findAllCurrencies() async {
+  Future<List<AccountCurrencyEntity>?> findAllCurrencies() async {
     return _queryAdapter.queryList('SELECT * FROM AccountCurrency',
-        mapper: (Map<String, dynamic> row) => AccountCurrencyEntity(
+        mapper: (Map<String, Object?> row) => AccountCurrencyEntity(
             accountcurrencyId: row['accountcurrency_id'] as String,
             accountId: row['account_id'] as String,
             currencyId: row['currency_id'] as String,
@@ -613,26 +612,18 @@ class _$AccountCurrencyDao extends AccountCurrencyDao {
   }
 
   @override
-  Future<AccountCurrencyEntity> findOneByAccountyId(String id) async {
+  Future<AccountCurrencyEntity?> findOneByAccountyId(String id) async {
     return _queryAdapter.query(
-        'SELECT * FROM AccountCurrency WHERE AccountCurrency.accountcurrency_id = ? LIMIT 1',
-        arguments: <dynamic>[id],
-        mapper: (Map<String, dynamic> row) => AccountCurrencyEntity(
-            accountcurrencyId: row['accountcurrency_id'] as String,
-            accountId: row['account_id'] as String,
-            currencyId: row['currency_id'] as String,
-            balance: row['balance'] as String,
-            numberOfUsedExternalKey: row['number_of_used_external_key'] as int,
-            numberOfUsedInternalKey: row['number_of_used_internal_key'] as int,
-            lastSyncTime: row['last_sync_time'] as int));
+        'SELECT * FROM AccountCurrency WHERE AccountCurrency.accountcurrency_id = ?1 LIMIT 1',
+        mapper: (Map<String, Object?> row) => AccountCurrencyEntity(accountcurrencyId: row['accountcurrency_id'] as String, accountId: row['account_id'] as String, currencyId: row['currency_id'] as String, balance: row['balance'] as String, numberOfUsedExternalKey: row['number_of_used_external_key'] as int, numberOfUsedInternalKey: row['number_of_used_internal_key'] as int, lastSyncTime: row['last_sync_time'] as int),
+        arguments: [id]);
   }
 
   @override
-  Future<List<JoinCurrency>> findJoinedByAccountId(String id) async {
+  Future<List<JoinCurrency>?> findJoinedByAccountId(String id) async {
     return _queryAdapter.queryList(
-        'SELECT * FROM JoinCurrency WHERE JoinCurrency.account_id = ?',
-        arguments: <dynamic>[id],
-        mapper: (Map<String, dynamic> row) => JoinCurrency(
+        'SELECT * FROM JoinCurrency WHERE JoinCurrency.account_id = ?1',
+        mapper: (Map<String, Object?> row) => JoinCurrency(
             accountcurrencyId: row['accountcurrency_id'] as String,
             currencyId: row['currency_id'] as String,
             symbol: row['symbol'] as String,
@@ -644,18 +635,18 @@ class _$AccountCurrencyDao extends AccountCurrencyDao {
             blockchainId: row['network_id'] as String,
             network: row['network'] as String,
             chainId: row['chain_id'] as int,
-            publish:
-                row['publish'] == null ? null : (row['publish'] as int) != 0,
-            contract: row['contract'] as String,
+            publish: (row['publish'] as int) != 0,
+            contract: row['contract'] as String?,
             decimals: row['decimals'] as int,
             type: row['type'] as String,
-            accountId: row['account_id'] as String));
+            accountId: row['account_id'] as String),
+        arguments: [id]);
   }
 
   @override
-  Future<List<JoinCurrency>> findAllJoinedCurrency() async {
+  Future<List<JoinCurrency>?> findAllJoinedCurrency() async {
     return _queryAdapter.queryList('SELECT * FROM JoinCurrency',
-        mapper: (Map<String, dynamic> row) => JoinCurrency(
+        mapper: (Map<String, Object?> row) => JoinCurrency(
             accountcurrencyId: row['accountcurrency_id'] as String,
             currencyId: row['currency_id'] as String,
             symbol: row['symbol'] as String,
@@ -667,9 +658,8 @@ class _$AccountCurrencyDao extends AccountCurrencyDao {
             blockchainId: row['network_id'] as String,
             network: row['network'] as String,
             chainId: row['chain_id'] as int,
-            publish:
-                row['publish'] == null ? null : (row['publish'] as int) != 0,
-            contract: row['contract'] as String,
+            publish: (row['publish'] as int) != 0,
+            contract: row['contract'] as String?,
             decimals: row['decimals'] as int,
             type: row['type'] as String,
             accountId: row['account_id'] as String));
@@ -694,7 +684,7 @@ class _$UtxoDao extends UtxoDao {
         _utxoEntityInsertionAdapter = InsertionAdapter(
             database,
             'Utxo',
-            (UtxoEntity item) => <String, dynamic>{
+            (UtxoEntity item) => <String, Object?>{
                   'utxo_id': item.utxoId,
                   'accountcurrency_id': item.accountcurrencyId,
                   'tx_id': item.txId,
@@ -705,7 +695,7 @@ class _$UtxoDao extends UtxoDao {
                   'key_index': item.keyIndex,
                   'script': item.script,
                   'timestamp': item.timestamp,
-                  'locked': item.locked == null ? null : (item.locked ? 1 : 0),
+                  'locked': item.locked ? 1 : 0,
                   'sequence': item.sequence,
                   'address': item.address
                 }),
@@ -713,7 +703,7 @@ class _$UtxoDao extends UtxoDao {
             database,
             'Utxo',
             ['utxo_id'],
-            (UtxoEntity item) => <String, dynamic>{
+            (UtxoEntity item) => <String, Object?>{
                   'utxo_id': item.utxoId,
                   'accountcurrency_id': item.accountcurrencyId,
                   'tx_id': item.txId,
@@ -724,7 +714,7 @@ class _$UtxoDao extends UtxoDao {
                   'key_index': item.keyIndex,
                   'script': item.script,
                   'timestamp': item.timestamp,
-                  'locked': item.locked == null ? null : (item.locked ? 1 : 0),
+                  'locked': item.locked ? 1 : 0,
                   'sequence': item.sequence,
                   'address': item.address
                 });
@@ -740,12 +730,10 @@ class _$UtxoDao extends UtxoDao {
   final UpdateAdapter<UtxoEntity> _utxoEntityUpdateAdapter;
 
   @override
-  Future<List<JoinUtxo>> findAllJoinedUtxosById(
-      String accountcurrencyId) async {
+  Future<List<JoinUtxo>?> findAllJoinedUtxosById(String id) async {
     return _queryAdapter.queryList(
-        'SELECT * FROM JoinUtxo WHERE JoinUtxo.accountcurrency_id = ?',
-        arguments: <dynamic>[accountcurrencyId],
-        mapper: (Map<String, dynamic> row) => JoinUtxo(
+        'SELECT * FROM JoinUtxo WHERE JoinUtxo.accountcurrency_id = ?1',
+        mapper: (Map<String, Object?> row) => JoinUtxo(
             row['utxo_id'] as String,
             row['accountcurrency_id'] as String,
             row['tx_id'] as String,
@@ -756,16 +744,17 @@ class _$UtxoDao extends UtxoDao {
             row['key_index'] as int,
             row['script'] as String,
             row['timestamp'] as int,
-            row['locked'] == null ? null : (row['locked'] as int) != 0,
+            (row['locked'] as int) != 0,
             row['sequence'] as int,
             row['address'] as String,
-            row['decimals'] as int));
+            row['decimals'] as int),
+        arguments: [id]);
   }
 
   @override
-  Future<List<JoinUtxo>> findAllJoinedUtxos() async {
+  Future<List<JoinUtxo>?> findAllJoinedUtxos() async {
     return _queryAdapter.queryList('SELECT * FROM JoinUtxo',
-        mapper: (Map<String, dynamic> row) => JoinUtxo(
+        mapper: (Map<String, Object?> row) => JoinUtxo(
             row['utxo_id'] as String,
             row['accountcurrency_id'] as String,
             row['tx_id'] as String,
@@ -776,16 +765,16 @@ class _$UtxoDao extends UtxoDao {
             row['key_index'] as int,
             row['script'] as String,
             row['timestamp'] as int,
-            row['locked'] == null ? null : (row['locked'] as int) != 0,
+            (row['locked'] as int) != 0,
             row['sequence'] as int,
             row['address'] as String,
             row['decimals'] as int));
   }
 
   @override
-  Future<List<UtxoEntity>> findAllUtxos() async {
+  Future<List<UtxoEntity>?> findAllUtxos() async {
     return _queryAdapter.queryList('SELECT * FROM Utxo',
-        mapper: (Map<String, dynamic> row) => UtxoEntity(
+        mapper: (Map<String, Object?> row) => UtxoEntity(
             row['utxo_id'] as String,
             row['accountcurrency_id'] as String,
             row['tx_id'] as String,
@@ -796,17 +785,16 @@ class _$UtxoDao extends UtxoDao {
             row['key_index'] as int,
             row['script'] as String,
             row['timestamp'] as int,
-            row['locked'] == null ? null : (row['locked'] as int) != 0,
+            (row['locked'] as int) != 0,
             row['address'] as String,
             row['sequence'] as int));
   }
 
   @override
-  Future<List<UtxoEntity>> findAllUtxosById(String accountcurrencyId) async {
+  Future<List<UtxoEntity>?> findAllUtxosById(String id) async {
     return _queryAdapter.queryList(
-        'SELECT * FROM Utxo WHERE Utxo.accountcurrency_id = ?',
-        arguments: <dynamic>[accountcurrencyId],
-        mapper: (Map<String, dynamic> row) => UtxoEntity(
+        'SELECT * FROM Utxo WHERE Utxo.accountcurrency_id = ?1',
+        mapper: (Map<String, Object?> row) => UtxoEntity(
             row['utxo_id'] as String,
             row['accountcurrency_id'] as String,
             row['tx_id'] as String,
@@ -817,17 +805,17 @@ class _$UtxoDao extends UtxoDao {
             row['key_index'] as int,
             row['script'] as String,
             row['timestamp'] as int,
-            row['locked'] == null ? null : (row['locked'] as int) != 0,
+            (row['locked'] as int) != 0,
             row['address'] as String,
-            row['sequence'] as int));
+            row['sequence'] as int),
+        arguments: [id]);
   }
 
   @override
-  Future<JoinUtxo> findJoinedUtxoById(String id) async {
+  Future<JoinUtxo?> findJoinedUtxoById(String id) async {
     return _queryAdapter.query(
-        'SELECT * FROM JoinUtxo WHERE JoinUtxo.utxo_id = ? limit 1',
-        arguments: <dynamic>[id],
-        mapper: (Map<String, dynamic> row) => JoinUtxo(
+        'SELECT * FROM JoinUtxo WHERE JoinUtxo.utxo_id = ?1 limit 1',
+        mapper: (Map<String, Object?> row) => JoinUtxo(
             row['utxo_id'] as String,
             row['accountcurrency_id'] as String,
             row['tx_id'] as String,
@@ -838,10 +826,11 @@ class _$UtxoDao extends UtxoDao {
             row['key_index'] as int,
             row['script'] as String,
             row['timestamp'] as int,
-            row['locked'] == null ? null : (row['locked'] as int) != 0,
+            (row['locked'] as int) != 0,
             row['sequence'] as int,
             row['address'] as String,
-            row['decimals'] as int));
+            row['decimals'] as int),
+        arguments: [id]);
   }
 
   @override
@@ -867,7 +856,7 @@ class _$ExchangeRateDao extends ExchangeRateDao {
         _exchangeRateEntityInsertionAdapter = InsertionAdapter(
             database,
             'ExchangeRate',
-            (ExchangeRateEntity item) => <String, dynamic>{
+            (ExchangeRateEntity item) => <String, Object?>{
                   'exchange_rate_id': item.exchangeRateId,
                   'name': item.name,
                   'rate': item.rate,
@@ -885,9 +874,9 @@ class _$ExchangeRateDao extends ExchangeRateDao {
       _exchangeRateEntityInsertionAdapter;
 
   @override
-  Future<List<ExchangeRateEntity>> findAllExchageRates() async {
+  Future<List<ExchangeRateEntity>?> findAllExchageRates() async {
     return _queryAdapter.queryList('SELECT * FROM ExchangeRate',
-        mapper: (Map<String, dynamic> row) => ExchangeRateEntity(
+        mapper: (Map<String, Object?> row) => ExchangeRateEntity(
             exchangeRateId: row['exchange_rate_id'] as String,
             name: row['name'] as String,
             rate: row['rate'] as String,
