@@ -5,10 +5,10 @@ import '../entity/account.dart';
 @dao
 abstract class AccountDao {
   @Query('SELECT * FROM Account')
-  Future<List<AccountEntity>> findAllAccounts();
+  Future<List<AccountEntity>?> findAllAccounts();
 
   @Query('SELECT * FROM Account WHERE account_id = :id LIMIT 1')
-  Future<AccountEntity> findAccount(String id);
+  Future<AccountEntity?> findAccount(String id);
 
   @Insert(onConflict: OnConflictStrategy.replace)
   Future<void> insertAccount(AccountEntity account);
