@@ -5,7 +5,7 @@ import '../entity/network.dart';
 @dao
 abstract class NetworkDao {
   @Query('SELECT * FROM Network')
-  Future<List<NetworkEntity>> findAllNetworks();
+  Future<List<NetworkEntity>?> findAllNetworks();
 
   @Insert(onConflict: OnConflictStrategy.replace)
   Future<List<int>> insertNetworks(List<NetworkEntity> networks);
