@@ -4,7 +4,7 @@ import '../entity/user.dart';
 @dao
 abstract class UserDao {
   @Query('SELECT * FROM User limit 1')
-  Future<UserEntity> findUser();
+  Future<UserEntity?> findUser();
 
   @Insert(onConflict: OnConflictStrategy.replace)
   Future<void> insertUser(UserEntity user);
@@ -14,5 +14,4 @@ abstract class UserDao {
 
   @delete
   Future<int> deleteUser(UserEntity user);
- 
 }
